Fix focusability of web suggest via talkback
The reasson is due to shortcut flow not having sent an accessibility
event. Shortcuts adding to home can use the same flow of how apps are
added to home by using bindInflatedItems() with an animation
endCallback(). This change should be fine since addInScreen() is called
in bindInflatedItems() too.

Bug:383461209
Test:Manually video- https://drive.google.com/file/d/1y3CslaIvRQMsODlQCYnGtG0Im4t7p4rR/view?usp=sharing
Flag: EXEMPT bug fix

Change-Id: I3b88b5ee40b4bcd43a83d362b6cbb0925120b700