Test all apps -> workspace drag for taskbar on home
Adds TAPL test that verifies users are able to drag app items from all
apps to workspace when all apps is opened on from a pinned taskbar shown
on the home screen, as may be the case when a desktop mode taskbar is
shown on home screen.

Adapts LauncherInstrumentation to handle the case where taskbar is shown
on home - when verifying launcher states, it sets taskbar visibility
expectations to account for a posibility that taskbar is visible in
home, or workspace states. Adds a request to the test protocol to
determine whether taskbar should be visible on home.

Bug: 392118517
Flag: com.android.window.flags.enter_desktop_by_default_on_freeform_displays
Test: Ran TaplTestLockedTaskbar tests
Change-Id: I40b06da70aaa49b0a3e0673797e6c18773a72238