Use FLAG_ALT_FOCUSABLE_IM in taskbar when needed
Previously taskbar would only set/remove the FLAG_NOT_FOCUSABLE, which
also made it IME focusable. This was toggled when a tasbkar folder was
opened/closed, as well as opening the popup by long pressing an app in
the taskbar. Folders do require IME focusability, as the folder name can
be edited. However, the popup does not require IME focusability, and
this lead to the IME suddenly disappearing when the popup is shown.

This fixes the issue by also setting FLAG_ALT_FOCUSABLE_IM as needed,
only when the taskbar is also focusable. This follows the same logic
as in NotificationShadeWindowControllerImpl.

Flag: EXEMPT bugfix
Test: switch to gesture nav, show IME in app, show popup and verify
  IME is still visible
  open folder in either taskbar or launcher and verify IME can be shown
Bug: 361046041
Change-Id: Idb9b4503d5622222e99735f0f6981a17add12360