Fix Navbar height issue
When used GMS load, the first time phone boot, the navbar height was abnormal during the setup wizard.
The reason is that when the phone boot, the height of the gesture layout is used.
Therefore, when using the three-buttons navigation layout, do not use the height of the gesture layout.

Bug: 397550729
Test: manual
      test device boot.

Change-Id: I6e7aaf279a410d03f6f0068f26a4e341673c4334
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>