Update NavHandle tests to use more conservative timings.
On some devices, the long press would trigger due to an overly
long wait in the test. Separately, the small slop tests were
using a slop that would trigger the second stage if that flag was
enabled, so the trigger didn't happen (WAI). We already test 2
stage, so I updated these tests to use a smaller slop that should
not trigger with or without 2 stage.

Test: atest NexusLauncherTests:com.android.quickstep.inputconsumers.NavHandleLongPressInputConsumerTest -- --abi arm64-v8a
Flag: EXEMPT tests
Fix: 398216382

Change-Id: Iec633d01eff63eac8ba072fef4ac1e7b452d8242