Not refresh predicted app in right click
When users right click the mouse/touchpad, predicted apps in hotseat refresh and it is weird.
It refreshes because the right click is treated as a drag and a drag would trigger predicted apps refresh.
This fix adds a PreDragCondition to avoid it from starting drag.

Bug: 343650193
Test: Manual, connect to touchpad and right click on widget and folderIcon
Flag: NONE, Bug fix
Change-Id: I650f34f02a920b61e3a1f1ada7d2da10e1625e47