Fix home screen page description when removing an existing page.
When we remove an in-between or last empty page, we manually set the
current page but don't update the description that's set on the
accessibility view. Due to which, when talkback focuses on it, it sees
the stale page count in the description.

This fix ensures to keep the state updated during removal.

Bug: 390360063
Fix: 390360063
Test: Talkback
Flag: EXEMPT BUGFIX
Change-Id: Idd9a54d3a1ee43e56f51db78e614776eca6a12f7