Skip TaskView settledProgress and fullscreenProgress updates for same progress value
This prevents calling settledProgress with the same value multiple times during the onFullscreenProgress animation. It was causing a performance issue when the icon visibility was switching between VISIBILE and INVISIBLE each frame of the animation. More information in the comments of b/398318613.

Bug: 398318613
Flag: com.android.launcher3.enable_overview_icon_menu
Test: Manual
Test: systemui-thermal-throttling-3-suite
Change-Id: Ib24bfd7541bb177d7141718fa36d2349f7ec9d46