Added logic to show the shell drop target and update the bubble location
This change leverages a bubble bar location parameter when an app icon
is dropped and displays the shell expanded view drop target when an icon
is dragged over the bubble bar drop zone.

Fixes: 397459664
Flag: com.android.wm.shell.enable_create_any_bubble
Test: Manual.
Drag a taskbar app icon to the same side as the bubble bar. Confirm that
the expanded view from shell is displayed.
Continue dragging to the opposite side of the bubble bar location drop
zone. Confirm that the expanded view drop target is displayed on the
opposite side.
Release the dragged icon in the drop zone. Confirm that the bubble bar
expands at the new location.

Change-Id: I956d66b81810abce074b10e012cf5c873003d84c