Fix TaskMenuView bottom row divider and translation.
Adjusts the positioning logic for the task menu when displayed for tasks on the bottom row of the grid. The changes ensure the menu is visible and does not overlap with other UI elements, such as the taskbar. It defines min and max bounds of the menu when it's opened on bottom row. It also moves the menu on top of the taskbar.

A follow-up CL will manage the max height to make the bottom row menu scrollable.

Fix: 322953749
Bug: 326952853
Flag: com.android.launcher3.enable_overview_icon_menu
Test: OverviewMenuImageTest
Change-Id: I44fb70d8537c0b52af4314e818d02ccfcd35d8d3