Handle bubble bar drag and bubble drag to fullscreen
This change wires bubble bar dragging as well as bubble dragging to
full screen behind a flag. Note that when the bubble is dragged to
full screen we remove the bubble from the bubble bar and notify
sysui but on the sysui side the bubble doesn't yet get converted.

Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Bug: 388858013
Test: manual
       - with flag off
         - verify existing dragging works
       - with flag on
         - verify bubble bar can be dragged across
         - verify bubble is removed when bubble dragged to full screen zone
Change-Id: I9e4d8ee3d2ba9dffff3dd4ba91dfce1a5c723d68