Fix Taskbar not present in Desktop Mode after unlocking
The problem: with taskbar animation for Desktop mode we were checking, if we are already in DW then don't recreate taskbar. This was put in as condition so where DisplayController Info change we don't recreate twice. first being from transilition listerners and second being form info change in display controller.

The solution: Ignore the info change listener when there is already ongoing recreation is in progress.

Test: Presubmit
Bug: 399826787
Flag: EXEMPT bugfix

Change-Id: Ib86e79b3b4c86e515e44d1d1dd7ca98ed694c365