Reland - Not refresh predicted app in right click
When users right click the mouse/touchpad, predicted apps in hotseat refresh and it is weird.
It refreshes because the right click is treated as a drag and a drag would trigger predicted apps refresh.
This fix adds a PreDragCondition when right click to avoid it from starting drag.

Bug: 343650193
Test: Manual, connect to touchpad and right click on widget and folderIcon
Flag: com.android.launcher3.remove_apps_refresh_on_right_click
Change-Id: Id0be87beb8a0e9baf78a4ba8ec4c93462ec7bb3e