Update KQS task view borders
Updates border color to have more contrast to the background, and adds
2dp padding between the border and the task content view - done by
updating BorderAnimator not to assume the border width and stroke are
the same.

Adding padding between the border and task view content exopsed an issue
where the second thumbnail in a split task's borders were not clipped
correctly when adding a border (it looked like the clip rect was not
properly scaled down). This is addressed by moving the tack view
background (and the outline) from the root KeyboardQuickSwitchTaskViews
to the content views (i.e. moving the outline to the view that does not
get scaled when adding a border, and having the second thumbnail end
edge of the clip rectangle match the clip rectangle of the clipped
ancestor view).

Bug: 398770098
Test: Manual
Flag: EXEMPT (a11y) bug fix

Change-Id: Ie058c932411a6b47815b0abf756f739c55b8342f