Fix back button touch region too small
The back button in 3-button-nav touch region was unintentionally
decreased with the predictive_back_three_button_nav flag. The reason is
that NearestTouchFrame did not consider the back button for its touch
forwarding logic, because the back button did not have isClickable set
to true. Manually setting the clickable attribute fixes that.

Bug: 393498908
Test: Manual, i.e. verified that touches close to back button are not routed to the home button
Flag: EXEMPT bugfix
Change-Id: I7ad1ee13f92aa947f1a86f8d65de79a9d9e13899