Updated animation to account bubble bar stash state.
Added logic to check the bubble bar stash state and perform appropriate
animations based on that state.

Bug: 399678274
Test: Manual: Have bubbles in the bubble bar. Go to an application.
Swipe up the taskbar handle so the taskbar is shown. Click the search
icon in the taskbar. Start a drag on any icon in the grid. The bubble
bar stashes immediately on drag start. When the icon is dragged over the
bubble bar drop zone, the bubble bar unstashes and is shown at the
dragged location.
Flag: com.android.wm.shell.enable_create_any_bubble

Change-Id: I9663a03a7daacdfaba071d5e35d5b7404cefed4e