Improve reliability of cleanup done in onDestroy.
As seen in the bug, an exception caused rest of onDestroy to not clean
up the other objects - which could cause other leaks.

So, for somewhat non-trivial clean ups, I've wrapped them in try-catch.
This way, with Log.e we still will know about them, but not crash
destruction of activity and other clean up.

Bug: 400793700
Flag: EXEMPT BUGFIX
Test: Leak canary and studio heap dump
Change-Id: I45e7298394eafff182a2c800eebfa5f772fbe2de