Ensure back button shows up in Simple View for SUW
This change adds support for the dynamic changing of navigation modes in
SUW such that the back button will be visible in both states. Previously
there was no way to switch navigation modes during Setup Wizard, before the
introduction of Simple View. The buttons were not being drawn correctly
(or at all).

Fix: 381016997
Fix: 381363853
Test: Rerun SUW, enable simple view and ensure that the back buttons
works in both orientations and that the SUW buttons are unaffected
Flag: EXEMPT bugfix

Change-Id: Ifbdf0c9cd06f1347e6c2c9a943711cefbf6bff2c