Update HomeSettings Font to Google Sans Flex
This will update everything in Launcher Home Settings except the At a Glance Dialog. That will be updated in a different CL. Also, the At A glance and Notifications settings landing pages are not owned by HomeSettings. These will be updated when the Settings app updates their fonts to Google Sans Flex, if that has not happened already.

Bug: 393616497
Test: Verified via hard-coded colors and more obvious fonts that these updates actually change the correct text views.
Flag: com.android.launcher3.gsf_res
Change-Id: I74f66760f0fdb486681a2db55df578c9bf50ef47