Remove flaky test from BubbleBarViewAnimatorTest
This test is flaky in robolectric but does not repro locally.
Looking at the test it verifies that the animation doesn't get
canceled immediately after it starts. So the scenario here is an
edge case for something that shouldn't happen generally because
the animation is controlled in this class, although theoretically
other classes could obtain a reference to the physics animator and
cancel it directly. But since the test flakes badly, we're better
off without it until it can be fixed.

Bug: 404773413
Flag: EXEMPT test only
Test: atest BubbleBarViewAnimatorTest
Change-Id: I24493b71842a2fe859ffe40b6f375ade7ee8535d